# swagger_docs.yaml
openapi: 3.0.2
info:
  title: StoryBoom AI API
  description: |
    API for StoryBoom AI - Case Study Generation Platform
    
    This API provides endpoints for:
    - User authentication and management
    - Case study creation and management
    - Interview management (provider and client)
    - Media generation (videos, podcasts)
    - Feedback collection and analysis
    - Metadata analysis
    
    ## Authentication
    Most endpoints require authentication. Use session-based authentication by logging in first.
    
    ## Rate Limits
    API calls are limited to prevent abuse. Contact support if you need higher limits.
    
    ## IP Access
    For production access, contact support to whitelist your IP address.
  version: 1.0.0
  contact:
    name: StoryBoom AI Support
    email: storyboomai@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:10000
    description: Development server
  - url: https://your-production-domain.com
    description: Production server
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Session-based authentication (login required)
...
paths:
...
  /api/metadata/analyze:
    post:
      tags:
        - Metadata
      summary: Analyze metadata
      description: Analyze metadata for a case study
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - case_study_id
              properties:
                case_study_id:
                  type: integer
      responses:
        '200':
          description: Metadata analysis completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  metadata:
                    type: object
  /api/metadata/regenerate/{case_study_id}:
    post:
      tags:
        - Metadata
      summary: Regenerate metadata
      description: Regenerate metadata for an existing case study
      parameters:
        - name: case_study_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Metadata regenerated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  metadata:
                    type: object
  
  # File download endpoints
...
  /api/sentiment_chart/{case_study_id}:
    get:
      tags:
        - Metadata
      summary: Get sentiment chart
      description: Retrieve sentiment analysis chart for a case study
      parameters:
        - name: case_study_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sentiment chart retrieved
          content:
            image/png:
              schema:
                type: string
                format: binary
...